+++ src/main/java/nextstep/subway/path/advice/PathAdvice.java
@@ -3,10 +3,10 @@ package nextstep.subway.path.advice;
 import nextstep.subway.line.exception.LineException;
 import nextstep.subway.path.ui.PathController;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
 
-@ControllerAdvice(assignableTypes = {PathController.class})
+@RestControllerAdvice(assignableTypes = {PathController.class})
 public class PathAdvice {
 
     @ExceptionHandler(LineException.class)
